backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
python-dotenv
psycopg2-binary
pydantic
python-dateutil
prophet
statsmodels
scikit-learn
numpy
